+++ CHANGELOG.md
@@ -25,8 +25,8 @@ NOTE 1: Requires MacOS v10.14 (Mojave) to build.
 * File format: generates Realms with format v10 (reads and upgrades file format v5 or later).
 
 ### Internal
-* Updated Realm Core from ??? to Realm Core v10.0.0-alpha.1.
-* Updated Realm Sync from ??? to Realm Sync v10.0.0-alpha.2.
+* Updated Realm Core from ??? to Realm Core v10.0.0-alpha.2.
+* Updated Realm Sync from ??? to Realm Sync v10.0.0-alpha.4.
 * Updated Realm Object Store to commit f5003ce365757bb2106ba4c317d98fd83b7f6c4e.